+++ src/app/auth/auth.module.ts
@@ -6,7 +6,7 @@ import { FormsModule, ReactiveFormsModule } from "@angular/forms";
 import { AngularFireDatabaseModule } from "angularfire2/database";
 import { AngularFireAuthModule } from "angularfire2/auth";
 import { AuthService } from "../services/auth.service";
-import { AuthComponent } from "./auth.component";
+import { AuthComponent } from "./auth/auth.component";
 import { AuthGuard } from "../services/auth-guard.service";
 import { MaterialComponentsModule } from "../material/material.module";
 import { config } from "../../../firebase.config";
